+++ front-end/src/components/Post/GovernanceSideBar/VoteRefrendum.tsx
@@ -5,10 +5,10 @@ import { /* Divider, */ Dropdown, DropdownProps, DropdownItemProps, Icon, Popup,
 import { ApiPromise } from '@polkadot/api';
 import { InjectedAccountWithMeta } from '@polkadot/extension-inject/types';
 
-import { Form } from '../../ui-components/Form';
-import Button from '../../ui-components/Button';
-import { NotificationContext } from '../../context/NotificationContext';
-import { NotificationStatus } from '../../types';
+import { Form } from '../../../ui-components/Form';
+import Button from '../../../ui-components/Button';
+import { NotificationContext } from '../../../context/NotificationContext';
+import { NotificationStatus } from '../../../types';
 
 type ConvictionType = 'Locked1x' | 'Locked2x' | 'Locked3x' | 'Locked4x' | 'Locked5x' | 'Locked6x';
