+++ front-end/src/components/Post/PostGovernanceInfo/PostReferendumInfo.tsx
@@ -2,28 +2,27 @@ import * as React from 'react';
 import { Grid } from 'semantic-ui-react';
 import styled from '@xstyled/styled-components';
 
-import AddressComponent from '../../components/Address';
-import { OnchainLinkProposalFragment } from '../../generated/graphql';
-import { chainProperties } from '../../global/networkConstants';
-import getNetwork from '../../util/getNetwork';
+import AddressComponent from '../../../ui-components/Address';
+import { OnchainLinkReferendumFragment } from '../../../generated/graphql';
 
 interface Props{
 	className?: string
-	onchainLink: OnchainLinkProposalFragment
+	onchainLink: OnchainLinkReferendumFragment
 }
 
-const currentNetwork = getNetwork();
-
-const PostProposalInfo = ({ className, onchainLink }: Props) => {
+const PostReferendumInfo = ({ className, onchainLink }: Props) => {
 	if (!onchainLink) return null;
 
 	const {
-		onchain_proposal: onchainProposal,
+		onchain_referendum: onchainReferendum,
 		proposer_address: proposerAddress
 	} = onchainLink;
-	const preimage = onchainProposal?.[0]?.preimage;
-	const depositAmount = onchainProposal?.[0]?.depositAmount;
 
+	if ( !onchainReferendum?.[0] ){
+		return null;
+	}
+
+	const { delay, end, preimage, voteThreshold } = onchainReferendum?.[0];
 	const { metaDescription, method, preimageArguments } = preimage || {};
 
 	return (
@@ -33,16 +32,30 @@ const PostProposalInfo = ({ className, onchainLink }: Props) => {
 				<Grid.Column mobile={16} tablet={8} computer={8}>
 					<div className='info_group'>
 						<h6>Proposer</h6>
-						<AddressComponent className='' address={proposerAddress} accountName={'Proposer Address'} />
+						<AddressComponent className='' address={proposerAddress} accountName={'Proposer Address'}/>
 					</div>
 				</Grid.Column>
-				{depositAmount && currentNetwork &&
-				<Grid.Column mobile={16} tablet={8} computer={8}>
-					<div className='info_group'>
-						<h6>Deposit</h6>
-						{parseInt(depositAmount) / Math.pow(10, chainProperties[currentNetwork].tokenDecimals) + ' ' + chainProperties[currentNetwork].tokenSymbol}
-					</div>
-				</Grid.Column>}
+				{delay &&
+					<Grid.Column mobile={16} tablet={8} computer={8}>
+						<div className='info_group'>
+							<h6>Delay</h6>
+							{delay}
+						</div>
+					</Grid.Column>}
+				{end &&
+					<Grid.Column mobile={16} tablet={8} computer={8}>
+						<div className='info_group'>
+							<h6>End</h6>
+							{end}
+						</div>
+					</Grid.Column>}
+				{voteThreshold &&
+					<Grid.Column mobile={16} tablet={8} computer={8}>
+						<div className='info_group'>
+							<h6>voteThreshold</h6>
+							{voteThreshold}
+						</div>
+					</Grid.Column>}
 				{method &&
 				<Grid.Row>
 					<Grid.Column mobile={16} tablet={8} computer={8}>
@@ -64,19 +77,19 @@ const PostProposalInfo = ({ className, onchainLink }: Props) => {
 							: null}
 					</Grid.Column>
 				</Grid.Row>}
+				{metaDescription &&
 				<Grid.Column mobile={16} tablet={16} computer={16}>
-					{ metaDescription &&
 					<div className='info_group'>
 						<h6>Description</h6>
 						{metaDescription}
-					</div>}
-				</Grid.Column>
+					</div>
+				</Grid.Column>}
 			</Grid>
 		</div>
 	);
 };
 
-export default styled(PostProposalInfo)`
+export default styled(PostReferendumInfo)`
 	background-color: white;
 	padding: 2rem 3rem 2rem 3rem;
 	border-style: solid;
